requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-bootstrap4==3.0.1
django-ckeditor==6.1.0
django-environ==0.4.5
django-js-asset==1.2.2
idna==3.2
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
selenium==3.141.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
